Typescriptify PublicitySyncer
